+++ src/wic/api/__init__.py
@@ -1,3 +1,3 @@
-from .api import Step, Workflow
+from .pythonapi import Step, Workflow
 
 __all__ = ["Step", "Workflow"]
